Serializing and Building the Form
Make the field serialized in app/models/post.rb.

Build out the form in app/views/posts/_form.html.erb.

We create an array of tags and iterate through them using each. For each tag we create a label and a checkbox. We name both post[tags][#{tag}] so that our values will be passed in the correct format to the controller, as well as so that when the label is clicked it will activate the checkbox. Finally we use @post.tags&.any?(tag) to set the checkboxes as initially checked if those values are already in our serialized store.

Inside app/controllers/posts_controller.rb we first add our fields to the parameter whitelist so they will get through from the form.

Create a new private method to format the data of the paramters how we want it, in this case we just want an array of the tag names stored in our serialized column.

Now update the create and update methods to use our formatting method.

Source: https://www.danielsellergren.com/serialization-and-check_box_tag-in-rails-6